backend/requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.9
django-stubs==5.2.8
django-stubs-ext==5.2.8
djangorestframework==3.16.1
djangorestframework-stubs==3.16.6
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
package-index==1.0.0rc0
packaging==25.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pyotp==2.9.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
serious==1.0.0.dev20
sqlparse==0.5.4
tomli==2.3.0
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
Pillow>=10.0.0
django-cors-headers
django-tenants

